+++ multiqc/modules/freyja/freyja.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python
-
 """ MultiQC module to parse output from Freyja """
 
 import logging