+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### MultiQC updates
 
+- Update Clipboard.JS ([#2396](https://github.com/MultiQC/MultiQC/pull/2396))
 - Remove Highcharts ([#2409](https://github.com/MultiQC/MultiQC/pull/2409))
 
 ### New modules
@@ -11,6 +12,7 @@
 ### Module updates
 
 - **BBTools**: fix: skip missing values in bbmap qahist ([#2411](https://github.com/MultiQC/MultiQC/pull/2411))
+- **Samtools**: support new `markdup` command ([#2254](https://github.com/MultiQC/MultiQC/pull/2254))
 
 ## [MultiQC v1.21](https://github.com/MultiQC/MultiQC/releases/tag/v1.21) - 2024-02-28